upgrade of db; backup conf; import man

+++ app/sql.py
@@ -11,7 +11,7 @@ from audite import track_changes
 from app.common import (
     log_create,
     read_json_dict,
-    tab_exists,
+    tab_exists_scheme,
     unpack_foreign,
 )
 from app.error import (
@@ -34,6 +34,37 @@ DB structure is described in ./conf/sql_scheme.json
 """
 
 
+def store_man_alternatives(man_alt: dict) -> None:
+    """store manufacturers into tables"""
+    man = pd.DataFrame(columns=[MAN_NAME, MAN_ALT_NAME])  # pyright: ignore
+    for k, val in man_alt.items():
+        for v in val:
+            new_row = pd.DataFrame(
+                {MAN_NAME: k, MAN_ALT_NAME: v}, index=[0]  # pyright: ignore
+            )
+            man = pd.concat([man, new_row], ignore_index=True)
+    put(dat=pd.DataFrame({MAN_NAME: man[MAN_NAME].unique()}), tab="MANUFACTURER")
+    base_man = getDF(
+            tab="MANUFACTURER",
+            search=list(man_alt.keys()),
+            where=[MAN_NAME]
+            )  # fmt: on
+    alt_man = pd.merge(left=man, right=base_man, how="left", on=MAN_NAME)
+    put(dat=alt_man, tab="ALTERNATIVE_MANUFACTURER")
+
+
+def get_man_alternatives() -> dict:
+    """get manufacturers from"""
+    alt_man={}
+    alt_man_df = getDF("ALTERNATIVE_MANUFACTURER", follow=True)
+    for m in alt_man_df[MAN_NAME].unique():
+        mask = alt_man_df[MAN_NAME] == m
+        alt_man[m]=alt_man_df.loc[mask,MAN_ALT_NAME].to_list()
+    return alt_man
+
+
+
+
 def put(dat: pd.DataFrame, tab: str, on_conflict: dict | None = None) -> Dict:
     """
     put DataFrame into sql at table=tab
@@ -41,7 +72,7 @@ def put(dat: pd.DataFrame, tab: str, on_conflict: dict | None = None) -> Dict:
     check if tab exists!
     put() may raise sql_executeError and sql_tabError
     """
-    tab_exists(tab)
+    tab_exists_scheme(tab)
     sql_scheme = read_json_dict(SQL_SCHEME)
 
     if dat.empty:
@@ -58,6 +89,8 @@ def put(dat: pd.DataFrame, tab: str, on_conflict: dict | None = None) -> Dict:
     sql_columns = tab_columns(tab)
     # take only columns applicable to table
     d = dat.loc[:, [c in sql_columns for c in dat.columns]]
+    if d.empty:
+        return {}
     return __write_table__(dat=d, tab=tab, on_conflict=on_conflict)  # pyright: ignore
 
 
@@ -78,20 +111,25 @@ def __write_table__(
     # Convert to list of tuples, which executemany handles correctly
     records = list(df_copy.itertuples(index=False, name=None))
 
+    # single element tuple add coma which brake column names
+    col_names = f"{list(dat.columns)}"
+    col_names = col_names.replace("[", "(")
+    col_names = col_names.replace("]", ")")
+
     action = on_conflict.get("action", "REPLACE")  # default is ON CONFLICT REPLACE
     if action == "UPDATE SET":
         unique_col = on_conflict["unique_col"]
         update_cols = ",".join(
             [f"{c} = {c} + EXCLUDED.{c}" for c in on_conflict["add_col"]]
         )
-        cmd = f"""INSERT INTO {tab} {tuple(dat.columns)}
+        cmd = f"""INSERT INTO {tab} {col_names}
                   VALUES ({','.join(['?'] * len(dat.columns))})
                   ON CONFLICT ({','.join(unique_col)})
                   DO UPDATE SET {update_cols}
                """
     else:
         action = "OR " + action
-        cmd = f"""INSERT {action} INTO {tab} {tuple(dat.columns)}
+        cmd = f"""INSERT {action} INTO {tab} {col_names}
                   VALUES ({','.join(['?'] * len(dat.columns))})
                """
 
@@ -250,7 +288,7 @@ def get(  # pylint: disable=too-many-branches
         follow: if True, search in all FOREIGN subtables
     """
     # check if tab exists!
-    tab_exists(tab)
+    tab_exists_scheme(tab)
     # assign values if default
     all_cols = tab_columns(tab)
     # normalize input to list
@@ -397,7 +435,7 @@ def sql_check() -> None:
     """Check db file if aligned with scheme written in sql_scheme.json.
     Check if table exists and if has the required columns.
     Creates one if necessery
-    DB location and name taken from ./conf/configuration/py
+    DB location and name taken from ./conf/config.py
     """
     # make sure if exists
     if not os.path.isfile(DB_FILE):
@@ -422,6 +460,9 @@ def sql_check() -> None:
         from_json_scheme = [str(c) for c in sql_scheme[tab].get("FOREIGN", [])]
         if sorted(from_sql_scheme) != sorted(from_json_scheme):
             raise SqlCheckError(DB_FILE, tab)
+    # check auditing tables (audite_changefeed)
+    if "audite_changefeed" not in sql_tables():
+        raise SqlCheckError(DB_FILE, "audite")
 
 
 def __sql_execute__(
@@ -482,14 +523,67 @@ def __sql_execute__(
         con.close()  # type: ignore
 
 
-def sql_create() -> None:  # pylint: disable=too-many-branches
+def __tab_cmd__(tab: str, col_def: Dict) -> str:
+    """
+    craft sql command to create table
+    """
+    tab_cmd = f"CREATE TABLE {tab} ("
+    for col in col_def:  # fmt: off
+        if col not in SQL_KEYWORDS:  # fmt: on
+            tab_cmd += f"{col} {col_def[col]}, "
+        elif col == "FOREIGN":  # FOREIGN
+            for foreign in col_def[col]:
+                k, v = list(foreign.items())[0]
+                # sqlite do not complain on missing foreign table
+                # during creation
+                try:
+                    _, f_table, _ = unpack_foreign(foreign)
+                    tab_exists_scheme(f_table)
+                except SqlTabError as err:
+                    msg.msg(str(err))
+                    raise SqlCreateError(SQL_SCHEME) from err
+
+                tab_cmd += f"FOREIGN KEY({k}) REFERENCES {v}, "
+    tab_cmd = re.sub(",[^,]*$", "", tab_cmd)  # remove last comma
+    tab_cmd += ") "
+    return tab_cmd
+
+
+def __check_scheme__(tab: str, col_def: Dict) -> None:
+    """
+    check correctness of sql scheme json file
+    """
+    # read_json already checks if each table is a dict
+    for col in col_def:
+        if col not in SQL_KEYWORDS:  # fmt: on
+            if not isinstance(col_def[col], str):
+                raise SqlSchemeError(tab=tab, key=col, expected="String")
+        elif col == "FOREIGN":  # FOREIGN
+            if not isinstance(col_def[col], List):
+                raise SqlSchemeError(tab=tab, key=col, expected="List")
+            for foreign in col_def[col]:
+                if not isinstance(foreign, dict):
+                    raise SqlSchemeError(tab=tab, key=col, expected="List of Dict")
+        elif col in ["UNIQUE", "HASH_COLS"]:
+            if not isinstance(col_def[col], List):
+                raise SqlSchemeError(tab=tab, key=col, expected="List")
+        elif col == "ON_CONFLICT":
+            if not isinstance(col_def[col], Dict):
+                raise SqlSchemeError(tab=tab, key=col, expected="Dict")
+        elif col == "COL_DESCRIPTION":
+            if not isinstance(col_def[col], Dict):
+                raise SqlSchemeError(tab=tab, key=col, expected="Dict")
+
+
+def sql_create(one_tab="") -> None:  # pylint: disable=too-many-branches
     """Creates sql query based on sql_scheme.json and send to db.
     Perform check if created DB is aligned with scheme from sql.json file.
+    if one_tab!="" create only one tab
     """
-    if os.path.isfile(DB_FILE):
+    if os.path.isfile(DB_FILE) and not one_tab:
         # just in case the file exists
+        # when creating only one tab assumption is that we want to add to existng db
         os.remove(DB_FILE)
-
     path = os.path.dirname(DB_FILE)
     if not os.path.isdir(path):
         raise CheckDirError(directory=path)
@@ -502,34 +596,16 @@ def sql_create() -> None:  # pylint: disable=too-many-branches
 
     # create tables query for db
     sql_cmd = []
+    if one_tab:
+        sql_scheme = {one_tab: sql_scheme[one_tab]}
     for tab in sql_scheme:
-        tab_cmd = f"CREATE TABLE {tab} ("
-        for col in sql_scheme[tab]:  # fmt: off
-            if not any(
-                [
-                    isinstance(sql_scheme[tab], dict),
-                    isinstance(sql_scheme[tab], list),
-                ]
-            ):
-                raise SqlSchemeError(tab=tab)
-            if col not in SQL_KEYWORDS:  # fmt: on
-                tab_cmd += f"{col} {sql_scheme[tab][col]}, "
-            elif col == "FOREIGN":  # FOREIGN
-                for foreign in sql_scheme[tab][col]:
-                    k, v = list(foreign.items())[0]
-                    # sqlite do not complain on missing foreign table
-                    # during creation
-                    try:
-                        _, f_table, _ = unpack_foreign(foreign)
-                        tab_exists(f_table)
-                    except SqlTabError as err:
-                        msg.msg(str(err))
-                        raise SqlCreateError(SQL_SCHEME) from err
+        try:
+            __check_scheme__(tab=tab, col_def=sql_scheme[tab])
+        except SqlSchemeError as err:
+            print(err)
+            raise SqlCreateError(SQL_SCHEME) from err
 
-                    tab_cmd += f"FOREIGN KEY({k}) REFERENCES {v}, "
-        tab_cmd = re.sub(",[^,]*$", "", tab_cmd)  # remove last comma
-        tab_cmd += ") "
-        sql_cmd.append(tab_cmd)
+        sql_cmd.append(__tab_cmd__(tab, sql_scheme[tab]))
         if unique_cols := str(
             sql_scheme[tab]["UNIQUE"] if "UNIQUE" in sql_scheme[tab].keys() else ""
         ):
@@ -549,20 +625,34 @@ def sql_create() -> None:  # pylint: disable=too-many-branches
     sql_cmd.append("SELECT tbl_name FROM sqlite_master WHERE type='table'")
 
     try:
-        status = __sql_execute__(sql_cmd)
+        __sql_execute__(sql_cmd)
     except SqlExecuteError as err:
         os.remove(DB_FILE)
         msg.msg(str(err))
         raise SqlCreateError(SQL_SCHEME) from err
 
-    if sorted(status[sql_cmd[-1][0:100]]["tbl_name"].to_list()) != sorted(
-        list(sql_scheme.keys())
-    ):
+    # check if all tables created
+    all_tables = sql_tables()
+    if not all(k in all_tables for k in sql_scheme.keys()):
         if os.path.isfile(DB_FILE):
             os.remove(DB_FILE)
         raise SqlCreateError(SQL_SCHEME)
+    # add auditing all changes on all tables
     for tab in sql_scheme:
         sql_audit(tab=tab)
+
+
+def sql_tables() -> List:
+    """
+    list all tables in sql db
+    """
+    sql_cmd = ["SELECT tbl_name FROM sqlite_master WHERE type='table'"]
+    try:
+        status = __sql_execute__(sql_cmd)
+        return status[sql_cmd[-1][0:100]]["tbl_name"].to_list()
+    except SqlExecuteError as err:
+        msg.msg(str(err))
+        return []
 
 
 def sql_audit(tab: str) -> None:
